Split Seed Service
