+++ MnestixSearcher.ApiServices/Contracts/ISearchService.cs
@@ -1,0 +1,12 @@
+﻿using MnestixSearcher.AasSearcher;
+using MongoDB.Driver;
+
+namespace MnestixSearcher.ApiServices.Contracts
+{
+    public interface ISearchService
+    {
+        Task<List<AasSearchEntry>> GetAsync();
+
+        Task<List<AasSearchEntry>> GetByCriteriaAsync(FilterDefinition<AasSearchEntry> filter);
+    }
+}
