+++ MnestixSearcher.ApiServices/Contracts/ISeedService.cs
@@ -1,14 +1,7 @@
-﻿using MnestixSearcher.AasSearcher;
-using MongoDB.Driver;
-
-namespace MnestixSearcher.ApiServices.Contracts
+﻿namespace MnestixSearcher.ApiServices.Contracts
 {
     public interface ISeedService
     {
-        Task<List<AasSearchEntry>> GetAsync();
-
-        Task<List<AasSearchEntry>> GetByCriteriaAsync(FilterDefinition<AasSearchEntry> filter);
-
         Task SeedDatabase();
     }
 }
