+++ MnestixSearcher.ApiServices/ServicesRegistration.cs
@@ -42,6 +42,7 @@ namespace MnestixSearcher.ApiServices
 
             services.AddTransient<IBase64Service, Base64Service>();
             services.AddScoped<ISeedService, SeedService>();
+            services.AddScoped<ISearchService, SearchService>();
             services.AddScoped<IAasService, AasService>();
             services.AddScoped<ISubmodelService, SubmodelService>();
             services.AddScoped<IFilterService, FilterService>();
