+++ MnestixSearcher/Schemes/Query/Query.cs
@@ -1,13 +1,12 @@
 ﻿using MnestixSearcher.AasSearcher;
 using MnestixSearcher.ApiServices.Contracts;
-using HotChocolate.Data;
 
 namespace MnestixSearcher.Schemes.Query
 {
     public class Query
     {
         [UseFiltering]
-        public async Task<List<AasSearchEntry>> GetEntries([Service] ISeedService aasSearcherService)
+        public async Task<List<AasSearchEntry>> GetEntries([Service] ISearchService aasSearcherService)
         {
             return await aasSearcherService.GetAsync();
         }
